requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
msal
msgraph-sdk
apscheduler
python-dotenv
Flask 